Revert "Fork the project"
