+++ .github/workflows/ci.yml
@@ -3,7 +3,7 @@ name: ci
 on: [push]
 
 jobs:
-  npm-test:
+  npm-test-jest:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
@@ -17,7 +17,7 @@ jobs:
 
       - uses: pnpm/action-setup@v2.2.4
         with:
-          version: 8.6.1
+          version: 7.14.2
           run_install: true
 
-      - run: pnpm --filter ./packages/pdf-snapshot-matcher test
+      - run: pnpm --filter ./packages/jest test
