+++ packages/jest/package.json
@@ -1,21 +1,18 @@
 {
-  "name": "@armandabric/pdf-snapshot-matcher",
+  "name": "@btravers/pdf-snapshot-jest",
   "version": "0.0.0",
-  "description": "PDF snapshot matcher",
+  "description": "Jest extension",
   "author": "Benoit TRAVERS",
   "license": "MIT",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "repository": {
     "type": "git",
-    "url": "https://github.com/armandabric/pdf-snapshot.git"
+    "url": "https://github.com/btravers/pdf-snapshot.git"
   },
   "keywords": [
     "test",
     "jest",
-    "expect",
-    "matcher",
-    "vitest",
     "snapshot",
     "pdf",
     "visual testing"
@@ -24,30 +21,29 @@
     "prebuild": "rimraf dist",
     "build": "tsc -p tsconfig.build.json",
     "prepare": "pnpm build",
-    "test": "jest",
-    "test:debug": "DEBUG=testcontainers* jest"
+    "test": "jest"
   },
   "dependencies": {
     "@trpc/client": "10.0.0-rc.6",
     "@trpc/server": "10.0.0-rc.6",
     "dedent": "^0.7.0",
     "lodash": "^4.17.21",
+    "node-fetch-commonjs": "^3.2.4",
     "rimraf": "^3.0.2"
   },
   "devDependencies": {
-    "@armandabric/pdf-snapshot-service": "workspace:^",
+    "@btravers/pdf-snapshot-service": "workspace:^",
     "@types/dedent": "^0.7.0",
     "@types/jest": "^29.2.2",
     "@types/lodash": "^4.14.188",
     "@types/node": "^18.11.9",
     "@types/rimraf": "^3.0.2",
-    "expect": "^29.5.0",
     "jest": "29.3.0",
     "testcontainers": "^9.0.0",
     "ts-jest": "^29.0.3",
     "typescript": "^4.8.4"
   },
   "peerDependencies": {
-    "expect": ">= 29"
+    "jest": ">= 28"
   }
 }
