+++ packages/jest/src/__pdf_snapshots__/01_jest-extension-should-perform-a-snapshot-test-of-a-dummy-pdf-snap.png
