+++ packages/jest/src/toMatchPdfSnapshot.spec.ts
@@ -1,9 +1,6 @@
 import * as path from 'node:path';
 import * as fs from 'node:fs';
 import { GenericContainer, StartedTestContainer } from 'testcontainers';
-import { toMatchPdfSnapshot } from './toMatchPdfSnapshot';
-
-expect.extend({ toMatchPdfSnapshot });
 
 describe('Jest extension', () => {
   let container: StartedTestContainer | undefined;
@@ -31,7 +28,6 @@ describe('Jest extension', () => {
     const pdfPath = path.join(__dirname, '__fixtures__/dummy.pdf');
     const pdf = await fs.promises.readFile(pdfPath);
 
-    // @ts-expect-error
     await expect(pdf).toMatchPdfSnapshot();
   });
 });
