+++ packages/service/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@armandabric/pdf-snapshot-service",
+  "name": "@btravers/pdf-snapshot-service",
   "version": "0.0.1",
   "description": "Main service",
   "author": "Benoit TRAVERS",
